api_metropole/requirements.txt
appdirs==1.4.4
arrow==1.3.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
botcity-crawler-plugin==0.2.2
botcity-framework-base==0.4.4
botcity-framework-core==1.2.1
botcity-framework-web==0.9.0
botcity-http-plugin==0.1.4
botcity-maestro-sdk==0.5.1
botcity-utils==0.1.2
bs4==0.0.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.7
cookiecutter==2.6.0
cssselect==1.2.0
distro==1.9.0
fake-useragent==1.5.1
idna==3.8
importlib_metadata==8.4.0
Jinja2==3.1.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msedge-selenium-tools==3.141.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parse==1.20.2
pillow==10.4.0
psutil==6.0.0
pyee==11.1.1
Pygments==2.18.0
pynput==1.7.7
pyperclip==1.9.0
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-xlib==0.33
pytz==2024.2
pywin32==306
pywinauto==0.6.8
PyYAML==6.0.2
requests==2.32.3
requests-html==0.10.0
requests-toolbelt==1.0.0
rich==13.8.1
selenium==3.141.0
six==1.16.0
soupsieve==2.6
text-unidecode==1.3
tqdm==4.66.5
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
w3lib==2.2.1
webdriver-manager==4.0.2
websockets==10.4
zipp==3.20.1
